chore(fix): Fix Merge changes not applied (#2181)
Changes done in #2167 and #2102 are not properly applied

- Set dynamic=false to old indices
- Fix the batch_size for data scan
