+++ src/argilla/server/daos/backend/client_adapters/elasticsearch.py
@@ -70,7 +70,7 @@ class ElasticsearchClient(OpenSearchClient):
 
         self.set_index_mappings(
             index=index,
-            mappings=vector_mappings,
+            properties=vector_mappings,
         )
 
     def reindex(
